Run code cleanup

+++ server/src/Application/Transactions/Commands/AddTransactionCommand.cs
@@ -64,11 +64,11 @@ public class AddTransactionCommandHandler : ICommandHandler<AddTransactionComman
             var paymentPeriod = new TimePeriod((DateOnly)request.PaymentStart!, (DateOnly)request.PaymentEnd!);
             var paymentFrequency = new Frequency(timeUnit, (int)request.TimesPerCycle!, (int)request.UnitsInCycle!);
             var paymentTimeline = new Timeline(paymentPeriod, paymentFrequency);
-            transaction = new Transaction((int)request.Profile!, request.Name, request.Description, 
+            transaction = new Transaction(request.Profile!, request.Name, request.Description,
                                           (decimal)request.Amount!, transactionType, paymentTimeline);
         } else {
             var timeline = new Timeline(new TimePeriod((DateOnly)request.PaymentStart!));
-            transaction = new Transaction((int)request.Profile!, request.Name, request.Description, 
+            transaction = new Transaction(request.Profile!, request.Name, request.Description,
                                           (decimal)request.Amount!, transactionType, timeline);
         }
